organize page query properly
derp
