+++ test/server/repositories/tests.js
@@ -79,7 +79,6 @@ lab.experiment("Tests Repository", function() {
 
         done();
       });
-
     });
 
     lab.test("returns an error when not enough rows inserted", function(done) {
@@ -93,7 +92,6 @@ lab.experiment("Tests Repository", function() {
 
         done();
       });
-
     });
 
     lab.test("returns an error when not enough rows inserted", function(done) {
@@ -106,37 +104,6 @@ lab.experiment("Tests Repository", function() {
 
         Code.expect(err).to.be.instanceof(Error);
         Code.expect(err.message).to.equal(testErrMsg);
-
-        done();
-      });
-
-    });
-
-  });
-
-  lab.experiment("sitemap", function() {
-    lab.test("returns an error when query fails", function(done) {
-      var testErrMsg = "testing";
-      var testErr = new Error(testErrMsg);
-
-      queryStub.callsArgWith(1, testErr);
-
-      tests.sitemap(function(err) {
-
-        Code.expect(err).to.be.instanceof(Error);
-        Code.expect(err.message).to.equal(testErrMsg);
-
-        done();
-      });
-    });
-
-    lab.test("returns tests to use for sitemap", function(done) {
-      queryStub.callsArgWith(1, null, []);
-
-      tests.sitemap(function(err, results) {
-
-        Code.expect(err).to.be.null();
-        Code.expect(results).to.be.instanceof(Array);
 
         done();
       });
